Make Graph and Multigraph fully internal
Fixes #20382
