--- src/EFCore/Internal/Graph.cs
@@ -1,71 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System.Collections.Generic;
-using JetBrains.Annotations;
-
-namespace Microsoft.EntityFrameworkCore.Internal
-{
-    /// <summary>
-    ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-    ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-    ///     any release. You should only use it directly in your code with extreme caution and knowing that
-    ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-    /// </summary>
-    public abstract class Graph<TVertex>
-    {
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public abstract IEnumerable<TVertex> Vertices { get; }
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public abstract IEnumerable<TVertex> GetOutgoingNeighbors([NotNull] TVertex from);
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public abstract IEnumerable<TVertex> GetIncomingNeighbors([NotNull] TVertex to);
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public virtual ISet<TVertex> GetUnreachableVertices([NotNull] IReadOnlyList<TVertex> roots)
-        {
-            var unreachableVertices = new HashSet<TVertex>(Vertices);
-            unreachableVertices.ExceptWith(roots);
-            var visitingQueue = new List<TVertex>(roots);
-
-            var currentVertexIndex = 0;
-            while (currentVertexIndex < visitingQueue.Count)
-            {
-                var currentVertex = visitingQueue[currentVertexIndex];
-                currentVertexIndex++;
-                // ReSharper disable once LoopCanBeConvertedToQuery
-                foreach (var neighbor in GetOutgoingNeighbors(currentVertex))
-                {
-                    if (unreachableVertices.Remove(neighbor))
-                    {
-                        visitingQueue.Add(neighbor);
-                    }
-                }
-            }
-
-            return unreachableVertices;
-        }
-    }
-}
