+++ src/EFCore/Metadata/Conventions/ModelCleanupConvention.cs
@@ -7,6 +7,7 @@ using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using Microsoft.EntityFrameworkCore.Metadata.Conventions.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
+using Microsoft.EntityFrameworkCore.Utilities;
 
 namespace Microsoft.EntityFrameworkCore.Metadata.Conventions
 {
@@ -30,7 +31,8 @@ namespace Microsoft.EntityFrameworkCore.Metadata.Conventions
         protected virtual ProviderConventionSetBuilderDependencies Dependencies { get; }
 
         /// <inheritdoc />
-        public virtual void ProcessModelFinalizing(IConventionModelBuilder modelBuilder, IConventionContext<IConventionModelBuilder> context)
+        public virtual void ProcessModelFinalizing(
+            IConventionModelBuilder modelBuilder, IConventionContext<IConventionModelBuilder> context)
         {
             RemoveEntityTypesUnreachableByNavigations(modelBuilder, context);
             RemoveNavigationlessForeignKeys(modelBuilder);
@@ -44,7 +46,7 @@ namespace Microsoft.EntityFrameworkCore.Metadata.Conventions
             var rootEntityTypes = GetRoots(model, ConfigurationSource.DataAnnotation);
             using (context.DelayConventions())
             {
-                foreach (var orphan in new ModelNavigationsGraphAdapter(model).GetUnreachableVertices(rootEntityTypes))
+                foreach (var orphan in new GraphAdapter(model).GetUnreachableVertices(rootEntityTypes))
                 {
                     modelBuilder.HasNoEntityType(orphan, fromDataAnnotation: true);
                 }
@@ -90,5 +92,28 @@ namespace Microsoft.EntityFrameworkCore.Metadata.Conventions
                 entityType.RemoveAnnotation(CoreAnnotationNames.NavigationCandidates);
             }
         }
+
+        private sealed class GraphAdapter : Graph<IConventionEntityType>
+        {
+            private readonly IConventionModel _model;
+
+            public GraphAdapter([NotNull] IConventionModel model)
+            {
+                _model = model;
+            }
+
+            public override IEnumerable<IConventionEntityType> Vertices => _model.GetEntityTypes();
+
+            public override IEnumerable<IConventionEntityType> GetOutgoingNeighbors(IConventionEntityType from)
+                => from.GetForeignKeys().Where(fk => fk.DependentToPrincipal != null).Select(fk => fk.PrincipalEntityType)
+                    .Union(
+                        from.GetReferencingForeignKeys().Where(fk => fk.PrincipalToDependent != null).Select(fk => fk.DeclaringEntityType));
+
+            public override IEnumerable<IConventionEntityType> GetIncomingNeighbors(IConventionEntityType to)
+                => to.GetForeignKeys().Where(fk => fk.PrincipalToDependent != null).Select(fk => fk.PrincipalEntityType)
+                    .Union(
+                        to.GetReferencingForeignKeys().Where(fk => fk.DependentToPrincipal != null).Select(fk => fk.DeclaringEntityType));
+        }
     }
 }
+
